Externalize functions
